aarch64: Export fe{enable,disable,get}except on Cygwin
For aarch64 on ELF targets, the library does not export
fe{enable,disable,get}except as symbols from the library, relying on
static inline functions to provide suitable definitions if required.
But for Cygwin we need to create real definitions to satisfy the DLL
export script.

So arrange for real definitions of these functions when building on Cygwin.

Signed-off-by: Radek Bartoň <radek.barton@microsoft.com>