Cygwin: spawn: apply PID_NEW_PG flag to spawned child
Previously, it was always applied to "myself", which is only appropriate
in the case of _P_OVERLAY (aka exec).  Apply the flag to "myself" only
in that case, and apply it to the new child instead in other cases.
Also, clear the flag from "myself" in the case of a failed exec.

Fixes: 8d8724ee1b5a ("Cygwin: pty: Fix Ctrl-C handling further for non-cygwin apps.")
Signed-off-by: Jeremy Drake <cygwin@jdrake.com>