amdgcn, libm: fix infinite loop
The end condition on this loop, unlike all the other similar loops, is
"i >= 0", which is a problem because "i <<= 1" can go negative and then zero if
you continue shifting, and so back to true, again.  This isn't a problem for
the loop in the scalar implementation, but it means we need to mask the shift
in the vector implementation.

This fixes GCC PR#121392.