Cygwin: fix finding overlaps from F_SETLKW.
The commit adding OFD locks changed from comparing just the F_POSIX and
F_LOCK flags to comparing the entire flags value in order to make sure
the locks are of the same type.  However, the F_WAIT flag may or may not
be set, and result in that comparison not matching.  Mask the F_WAIT
flag when attempting to compare the types of the locks.

This fixes the "many_locks" stc.

Signed-off-by: Jeremy Drake <cygwin@jdrake.com>
Fixes: a66ed519884d ("Cygwin: fcntl: implement Open File Description (OFD) locks")