amdgcn, libm: fix vector ilogb bugs (bug 33272)
The vector ilogb routines, including the ones inlined into fmod, had a bug
in which the conditional masks were not properly applied, causing the value of
one lane to be affected by conditional choices of another lane.  The problem
was not immediately obviously because all values were calculated correctly when
no lane contained a subnormal input.

The problem is fixed by proper use of VECTOR_COND_MOVE and VECTOR_WHILE.