Cygwin: Add a configure-time check for minimum w32api headers version
Since we now require w32api-headers >= 13 for the
AllocConsoleWithOptions() prototype and flags, add a configure-time
check for that, as I've mused about a couple of times before.

This maybe gives a more obvious diagnosis of the problem than 'not
declared' errors, and is perhaps more self-documenting about our
expectations here.

Future work: Most of the other conditionals on __MINGW64_VERSION_MAJOR
can probably be removed.