Cygwin: uinfo: allow to override user account as primary group
Do not only allow to override the (localized) group "None" as primary
group, but also the user account.  The user account is used as primary
group in the user token, if the user account is a Microsoft Account or
an AzureAD account.

Fixes: dc7b67316d01 ("Cygwin: uinfo: prefer token primary group")
Signed-off-by: Corinna Vinschen <corinna@vinschen.de>
(cherry picked from commit 351adcd96f026a4ac4e771764d5dbc512190dbf2)