Cygwin: uinfo: fix overriding group from SAM comment on AD member machines
When overriding the (localized) primary group "None" of a local SAM
account via SAM comment entry (e.g. '<cygwin group="some_group"/>') on a
Active Directory domain member machine, we have to take into account,
that the local account domain (actually the machine name) is always
prepended to local account names, i. e.

  MACHINE+account

because the names without prepended domain are reserved for the
primary AD domain accounts.

Therefore commit cc332c9e271b added code to prepend the local account
domain to the group name from the SAM comment, if the machine is a
domain member.

But here's the problem:

If the group in the SAM comment entry is a real local group, prepending
the local account domain is all nice and dandy.  But if the account used
in the SAM comment is a builtin like "Authenticated Users" (S-1-5-11) or
an alias like "Users" (S-1-5-32-545), this falls flat.

This patch keeps the check for "MACHINE+account" first.  This avoids
fetching the AD group rather than the local SAM group, if a local
group has the same name as an AD group.

But now, if the group prepended with the local account domain doesn't
result in a valid group entry, try again with the naked group name, to
allow aliases or builtin accounts to pass as primary group.

Fixes: cc332c9e271b ("* uinfo.cc [...] (pwdgrp::fetch_account_from_windows): Drop outdated comment.  Fix code fetching primary group gid of group setting in SAM description field.")
Signed-off-by: Corinna Vinschen <corinna@vinschen.de>
(cherry picked from commit 0870013a6487a25c1dc24f255daa0e6afd396bfb)