Cygwin: pty: Fix state initialization in pty_master_fwd_thread
If the ESC sequence ends at the middle of the sequence, the 'state'
remains non-zero. This prevent the next state machine from working
as expected. On contrary, 'start_at' is set in each parser, so is
not necessary to be initialized.

Fixes: 10d083c745dd ("Cygwin: pty: Inherit typeahead data between two input pipes.")
Suggested-by: Johannes Schindelin <Johannes.Schindelin@gmx.de>
Signed-off-by: Takashi Yano <takashi.yano@nifty.ne.jp>
(cherry picked from commit 53bbf4756c7f7c4e300406cd7f33a876ca9b4731)